pyproject.toml
[project]
name = "rutineret"
version = "0.1.0"
description = "Django REST API with JWT authentication and role-based access control"
authors = [{name = "Rutineret Team"}]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django==5.2.10",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "django-cors-headers==4.9.0",
    "drf-spectacular>=0.27.0",
    "Pillow==12.1.0",
    "django-allauth==65.4.0",
    "django-oauth-toolkit==2.4.0",
    "requests==2.32.5",
    "celery==5.4.0",
    "redis==5.2.1",
    "django-celery-beat==2.8.1",
    "django-celery-results==2.6.0",
    "pytz==2024.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "ruff",
    "poethepoet",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "ruff",
    "poethepoet",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "users"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "api.settings.development"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--cov=users --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 88
target-version = "py314"
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "DJ"]
lint.exclude = [
    ".venv",
    "migrations",
    "__pycache__",
    ".ruff_cache",
]

[tool.poe.tasks]
dev = "manage.py runserver --settings=api.settings.development"
test = "pytest --cov=users --cov-report=term-missing"
test-verbose = "pytest -v --cov=users --cov-report=term-missing --cov-report=html"
migrate = "manage.py migrate"
setup = "manage.py setup_initial_data"
shell = "manage.py shell"
shell-plus = "manage.py shell_plus"
reset-db = "manage.py flush --noinput"
collectstatic = "manage.py collectstatic --noinput"
createsuperuser = "manage.py createsuperuser"
makemigrations = "manage.py makemigrations"
check = "ruff check ."
format = "ruff format ."
fix = "ruff check --fix ."
lint = "ruff check . && ruff format --check ."
